Add basic smoke tests for deployment
To ensure that we catch any issues with our infrastructure and provisioning
management, we have implemented a very brief smoke test hook, so we can
detect any issue we have with our deployment.

These are currently based on a simple cURL check of the resulting web
urls http status codes, however we should move to a wider test tool
at some stage.